move asset contacts libs to ui.libs profileUtils
- refactor ui.libs eslint
- [CORE-959]
