--- core/ui.libs/src/index.html
@@ -1,30 +1,0 @@
-<div class="evan theme-evan">
-  <div class="container-flex border p-3 m-3 border bg-light">
-    <h3 class="mb-3">Buttons</h3>
-
-    <button type="button" class="btn btn-rounded btn-primary font-weight-normal">
-      Primary
-      <i class="fas fa-arrow-right label ml-2"></i>
-    </button>
-
-    <button type="button" class="btn btn-primary btn-circle">
-      <i class="fas fa-plus"></i>
-    </button>
-
-    <button type="button" class="btn btn-secondary btn-circle">
-      <i class="fas fa-pen"></i>
-    </button>
-
-    <button type="button" class="btn btn-outline-secondary btn-circle">
-      <i class="fas fa-share-alt"></i>
-    </button>
-  </div>
-
-  <div class="container-flex border p-3 m-3 border bg-light">
-    <h3 class="mb-3">
-      <a href="https://fontawesome.com" target="_blank">Fontawesome</a>
-    </h3>
-    
-    <iframe src="https://fontawesome.com/icons?d=gallery&q=share&m=free" width="100%" height="500px"></iframe>
-  </div>
-</div>
